Rename file UploadImageForm to ProfileImageForm
